+++ src/Smartstore.Core/Platform/Identity/Domain/SystemCustomerNames.cs
@@ -2,7 +2,7 @@
 {
     public static partial class SystemCustomerNames
     {
-        public static string SearchEngine => "SearchEngine";
+        public static string Bot => "SearchEngine";
         public static string BackgroundTask => "BackgroundTask";
         public static string PdfConverter => "PdfConverter";
         public static string WebhookClient => "WebhookClient";