+++ src/Smartstore.Web/Areas/Admin/Controllers/ActivityLogController.cs
@@ -134,7 +134,7 @@ namespace Smartstore.Admin.Controllers
 
                 if (systemCustomer != null)
                 {
-                    if (systemCustomer.IsSearchEngineAccount())
+                    if (systemCustomer.IsBot())
                     {
                         model.SystemAccountName = T("Admin.System.SystemCustomerNames.SearchEngine");
                     }