+++ changelog.md
@@ -19,6 +19,7 @@
 - Logout was not possible after new customer registrations.
 - Fixed unpublished products should be assignable as "promoted" to a product.
 - UrlRewriter: raw rules were not loaded from legacy storage
+- Fixed a topic was not editable if a menu link with an invalid target was associated.
 
 
 ## Smartstore 5.0.4